pyproject.toml
[tool.poetry]
name = "link-tracker"
version = "0.1.0"
description = ""
authors = ["Alexander Dmitriev <alexander.dmitriev2580@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
# sqlalchemy = "^1.4.23"
sqlalchemy = ">=1.4,<1.4.23"
psycopg2 = "^2.9.1"
Jinja2 = "^3.0.1"
python-multipart = "^0.0.5"
alembic = "^1.6.5"
google-auth = "^2.0.2"
requests = "^2.26.0"
google-api-python-client = "^2.19.1"
google-auth-oauthlib = "^0.4.6"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
